Switch to bytes analysis and other optimizations
The basic trick here is to find the index of the column divider once,
and then assume all other output will match it.  We use a subslice
operator to pull the key and value out.

BenchmarkParseStatus	    1000	   2270344 ns/op
BenchmarkParseVariables	    1000	   1503680 ns/op
BenchmarkConvert	 5000000	       354 ns/op

This is very close to PCRE's result, but the code is more complex and I
suspect a bit more breakable.

The advantage to this code is it has no external libraries, so it
should run on any supported go platform (with or without PCRE).  It
also makes it easier to build for go n00bs.

I'm on the fence as to which is better, but I'll leave this alone for
now.