Fix issue #16
A few main problems:
1) did not send an initial SHOW GLOBAL STATUS to mysql on startup, but
waited 1 interval (easy fix).  This accounted for the first missed
interval.
2) Output from the MYSQLCLI was not terminated, so the parser couldn't
detect a sample break until the next sample came in.  Fixed this by
adding a SELECT 'END' after the command and breaking on that.  Get rid
of the column headers altogether in the mysql batch output.
3) The output from the mysql command was buffered, so added a CLI
option to turn this off.